+++ packages/api-elasticsearch-tasks/tsconfig.build.json
@@ -4,7 +4,7 @@
   "references": [
     { "path": "../api/tsconfig.build.json" },
     { "path": "../api-core/tsconfig.build.json" },
-    { "path": "../api-dynamodb-to-elasticsearch/tsconfig.build.json" },
+    { "path": "../api-sync-to-opensearch/tsconfig.build.json" },
     { "path": "../api-opensearch/tsconfig.build.json" },
     { "path": "../aws-sdk/tsconfig.build.json" },
     { "path": "../background-tasks/tsconfig.build.json" },
@@ -28,8 +28,8 @@
       "@webiny/api": ["../api/src"],
       "@webiny/api-core/*": ["../api-core/src/*"],
       "@webiny/api-core": ["../api-core/src"],
-      "@webiny/api-dynamodb-to-elasticsearch/*": ["../api-dynamodb-to-elasticsearch/src/*"],
-      "@webiny/api-dynamodb-to-elasticsearch": ["../api-dynamodb-to-elasticsearch/src"],
+      "@webiny/api-sync-to-opensearch/*": ["../api-sync-to-opensearch/src/*"],
+      "@webiny/api-sync-to-opensearch": ["../api-sync-to-opensearch/src"],
       "@webiny/api-opensearch/*": ["../api-opensearch/src/*"],
       "@webiny/api-opensearch": ["../api-opensearch/src"],
       "@webiny/aws-sdk/*": ["../aws-sdk/src/*"],
