+++ packages/api-elasticsearch-tasks/tsconfig.json
@@ -4,7 +4,7 @@
   "references": [
     { "path": "../api" },
     { "path": "../api-core" },
-    { "path": "../api-dynamodb-to-elasticsearch" },
+    { "path": "../api-sync-to-opensearch" },
     { "path": "../api-opensearch" },
     { "path": "../aws-sdk" },
     { "path": "../background-tasks" },
@@ -28,8 +28,8 @@
       "@webiny/api": ["../api/src"],
       "@webiny/api-core/*": ["../api-core/src/*"],
       "@webiny/api-core": ["../api-core/src"],
-      "@webiny/api-dynamodb-to-elasticsearch/*": ["../api-dynamodb-to-elasticsearch/src/*"],
-      "@webiny/api-dynamodb-to-elasticsearch": ["../api-dynamodb-to-elasticsearch/src"],
+      "@webiny/api-sync-to-opensearch/*": ["../api-sync-to-opensearch/src/*"],
+      "@webiny/api-sync-to-opensearch": ["../api-sync-to-opensearch/src"],
       "@webiny/api-opensearch/*": ["../api-opensearch/src/*"],
       "@webiny/api-opensearch": ["../api-opensearch/src"],
       "@webiny/aws-sdk/*": ["../aws-sdk/src/*"],
