+++ packages/api-headless-cms-ddb-es/__tests__/__api__/setupFile.js
@@ -14,7 +14,7 @@ import {
 import { getBaseConfiguration } from "@webiny/api-opensearch";
 import { OpenSearchClient } from "@webiny/api-opensearch/exports/api/opensearch.js";
 import { getOpenSearchIndexPrefix } from "@webiny/api-opensearch";
-import { createDdbToEsStreamHandler } from "@webiny/api-dynamodb-to-elasticsearch";
+import { createDdbToEsStreamHandler } from "@webiny/api-sync-to-opensearch";
 
 if (typeof registerCmsOpenSearchStorageOperations !== "function") {
     throw new Error(`Loaded plugins file must export a function that returns an array of plugins.`);
