+++ packages/api-headless-cms-ddb-es/tsconfig.build.json
@@ -14,7 +14,7 @@
     { "path": "../utils/tsconfig.build.json" },
     { "path": "../api-core/tsconfig.build.json" },
     { "path": "../api-core-testing/tsconfig.build.json" },
-    { "path": "../api-dynamodb-to-elasticsearch/tsconfig.build.json" },
+    { "path": "../api-sync-to-opensearch/tsconfig.build.json" },
     { "path": "../background-tasks/tsconfig.build.json" },
     { "path": "../event-handler-core/tsconfig.build.json" },
     { "path": "../handler/tsconfig.build.json" },
@@ -52,8 +52,8 @@
       "@webiny/api-core": ["../api-core/src"],
       "@webiny/api-core-testing/*": ["../api-core-testing/src/*"],
       "@webiny/api-core-testing": ["../api-core-testing/src"],
-      "@webiny/api-dynamodb-to-elasticsearch/*": ["../api-dynamodb-to-elasticsearch/src/*"],
-      "@webiny/api-dynamodb-to-elasticsearch": ["../api-dynamodb-to-elasticsearch/src"],
+      "@webiny/api-sync-to-opensearch/*": ["../api-sync-to-opensearch/src/*"],
+      "@webiny/api-sync-to-opensearch": ["../api-sync-to-opensearch/src"],
       "@webiny/background-tasks/*": ["../background-tasks/src/*"],
       "@webiny/background-tasks": ["../background-tasks/src"],
       "@webiny/event-handler-core/*": ["../event-handler-core/src/*"],
