+++ packages/api-headless-cms-ddb-es/tsconfig.json
@@ -14,7 +14,7 @@
     { "path": "../utils" },
     { "path": "../api-core" },
     { "path": "../api-core-testing" },
-    { "path": "../api-dynamodb-to-elasticsearch" },
+    { "path": "../api-sync-to-opensearch" },
     { "path": "../background-tasks" },
     { "path": "../event-handler-core" },
     { "path": "../handler" },
@@ -52,8 +52,8 @@
       "@webiny/api-core": ["../api-core/src"],
       "@webiny/api-core-testing/*": ["../api-core-testing/src/*"],
       "@webiny/api-core-testing": ["../api-core-testing/src"],
-      "@webiny/api-dynamodb-to-elasticsearch/*": ["../api-dynamodb-to-elasticsearch/src/*"],
-      "@webiny/api-dynamodb-to-elasticsearch": ["../api-dynamodb-to-elasticsearch/src"],
+      "@webiny/api-sync-to-opensearch/*": ["../api-sync-to-opensearch/src/*"],
+      "@webiny/api-sync-to-opensearch": ["../api-sync-to-opensearch/src"],
       "@webiny/background-tasks/*": ["../background-tasks/src/*"],
       "@webiny/background-tasks": ["../background-tasks/src"],
       "@webiny/event-handler-core/*": ["../event-handler-core/src/*"],
