+++ packages/api-sync-to-opensearch/__tests__/event.test.ts
