+++ packages/api-sync-to-opensearch/ci.config.json
