+++ packages/api-sync-to-opensearch/src/DdbToEsLambdaHandler.ts
