+++ packages/api-sync-to-opensearch/src/NotEnoughRemainingTimeError.ts
