+++ packages/api-sync-to-opensearch/src/Operations.ts
