+++ packages/api-sync-to-opensearch/src/OperationsBuilder.ts
