+++ packages/api-sync-to-opensearch/src/SynchronizationBuilder.ts
