+++ packages/api-sync-to-opensearch/src/createDdbToEsStreamHandler.ts
