+++ packages/api-sync-to-opensearch/src/helpers/getNumberEnvVariable.ts
