+++ packages/api-sync-to-opensearch/src/helpers/shouldShowLogs.ts
