+++ packages/api-sync-to-opensearch/src/index.ts
