+++ packages/api-sync-to-opensearch/src/types.ts
