+++ packages/api-sync-to-opensearch/tsconfig.build.json
