+++ packages/api-sync-to-opensearch/tsconfig.json
