+++ packages/api-sync-to-opensearch/webiny.config.js
