+++ packages/project-aws-template/package.json
@@ -19,7 +19,6 @@
     "@webiny/api-background-tasks-os": "0.0.0",
     "@webiny/api-core-ddb": "0.0.0",
     "@webiny/api-core-sql": "0.0.0",
-    "@webiny/api-dynamodb-to-elasticsearch": "0.0.0",
     "@webiny/api-event-handler-aws": "0.0.0",
     "@webiny/api-event-handler-aws-ddb": "0.0.0",
     "@webiny/api-event-handler-aws-ddb-os": "0.0.0",
@@ -41,6 +40,7 @@
     "@webiny/api-scheduler": "0.0.0",
     "@webiny/api-scheduler-aws": "0.0.0",
     "@webiny/api-scheduler-server": "0.0.0",
+    "@webiny/api-sync-to-opensearch": "0.0.0",
     "@webiny/api-website-builder": "0.0.0",
     "@webiny/api-website-builder-workflows": "0.0.0",
     "@webiny/api-websockets": "0.0.0",
@@ -76,7 +76,7 @@
         "@webiny/api-background-tasks-os",
         "@webiny/api-core-ddb",
         "@webiny/api-core-sql",
-        "@webiny/api-dynamodb-to-elasticsearch",
+        "@webiny/api-sync-to-opensearch",
         "@webiny/api-event-handler-aws",
         "@webiny/api-event-handler-aws-ddb",
         "@webiny/api-event-handler-aws-ddb-os",
